Use uint8_t for statusBitset in CDKGDebugMemberStatus and CDKGDebugSessionStatus
No need to waste one byte per member and per LLMQ type.

+++ src/llmq/quorums_debug.h
@@ -36,7 +36,7 @@ public:
             bool receivedJustification : 1;
             bool receivedPrematureCommitment : 1;
         };
-        uint16_t statusBitset;
+        uint8_t statusBitset;
     };
 
     std::set<uint16_t> complaintsFromMembers;
@@ -74,7 +74,7 @@ public:
 
             bool aborted : 1;
         };
-        uint16_t statusBitset;
+        uint8_t statusBitset;
     };
 
     std::vector<CDKGDebugMemberStatus> members;
